refactor: response 관련 변환 로직을 LibraryStockDataMapper의 static 메소드로 분리
