+++ src/main/java/com/bookspot/stock/application/LibraryStockRefreshService.java
@@ -1,11 +1,11 @@
 package com.bookspot.stock.application;
 
 import com.bookspot.global.DateHolder;
+import com.bookspot.stock.application.mapper.LibraryStockDataMapper;
 import com.bookspot.stock.domain.LibraryStock;
 import com.bookspot.stock.domain.LibraryStockRepository;
 import com.bookspot.stock.domain.LoanState;
 import com.bookspot.stock.domain.LoanStateRefreshService;
-import com.bookspot.stock.presentation.response.LoanStateResponseEnum;
 import com.bookspot.stock.presentation.response.StockLoanStateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,7 +28,7 @@ public class LibraryStockRefreshService {
             return new StockLoanStateResponse(
                     stock.getId(),
                     stock.getUpdatedAt(),
-                    toDto(stock.getLoanState())
+                    LibraryStockDataMapper.transform(stock.getLoanState())
             );
 
         // TODO: 락을 다루는 일이기 떄문에 batch 처리와 충돌이 있을 수 있음. 관리 필요. 그래도 only 레코드 락이라...
@@ -45,16 +45,7 @@ public class LibraryStockRefreshService {
         return new StockLoanStateResponse(
                 stock.getId(),
                 now,
-                toDto(result)
+                LibraryStockDataMapper.transform(result)
         );
-    }
-
-    public LoanStateResponseEnum toDto(LoanState loanState) {
-        return switch (loanState) {
-            case LOANABLE -> LoanStateResponseEnum.LOANABLE;
-            case ON_LOAN -> LoanStateResponseEnum.ON_LOAN;
-            case UNKNOWN -> LoanStateResponseEnum.UNKNOWN;
-            case ERROR -> LoanStateResponseEnum.ERROR;
-        };
     }
 }
