+++ src/main/java/com/bookspot/stock/application/LibraryStockService.java
@@ -1,10 +1,8 @@
 package com.bookspot.stock.application;
 
+import com.bookspot.stock.application.mapper.LibraryStockDataMapper;
 import com.bookspot.stock.domain.LibraryStock;
 import com.bookspot.stock.domain.LibraryStockRepository;
-import com.bookspot.stock.domain.LoanState;
-import com.bookspot.stock.presentation.response.LoanStateResponseEnum;
-import com.bookspot.stock.presentation.response.StockLoanStateResponse;
 import com.bookspot.stock.presentation.response.StockLoanStateResponseList;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -37,25 +35,8 @@ public class LibraryStockService {
 
         return new StockLoanStateResponseList(
                 stocks.stream()
-                        .map(LibraryStockService::transform)
+                        .map(LibraryStockDataMapper::transform)
                         .toList()
         );
-    }
-
-    public static StockLoanStateResponse transform(LibraryStock stock) {
-        return new StockLoanStateResponse(
-                stock.getId(),
-                stock.getUpdatedAt(),
-                transform(stock.getLoanState())
-        );
-    }
-
-    public static LoanStateResponseEnum transform(LoanState loanState) {
-        return switch (loanState) {
-            case LOANABLE -> LoanStateResponseEnum.LOANABLE;
-            case ON_LOAN -> LoanStateResponseEnum.ON_LOAN;
-            case UNKNOWN -> LoanStateResponseEnum.UNKNOWN;
-            case ERROR -> LoanStateResponseEnum.ERROR;
-        };
     }
 }
